Move _host_by_id_cache and _evict_cached_host to BaseProviderInstance
Problem: _host_by_id_cache (PrivateAttr) and _evict_cached_host were
identically copy-pasted into all four provider instance classes
(Docker, Lima, Modal, VpsDocker), which all inherit from BaseProviderInstance.

Fix: Consolidated the shared cache attribute and eviction method into
BaseProviderInstance so the logic lives in one place.
