+++ libs/mngr_lima/imbue/mngr_lima/instance.py
@@ -92,16 +92,7 @@ class LimaProviderInstance(BaseProviderInstance):
 
     config: LimaProviderConfig = Field(frozen=True, description="Lima provider configuration")
 
-    _host_by_id_cache: dict[HostId, HostInterface] = PrivateAttr(default_factory=dict)
     _lima_checked: bool = PrivateAttr(default=False)
-
-    def _evict_cached_host(self, host_id: HostId, replacement: HostInterface | None = None) -> None:
-        """Remove a Host from the cache, disconnecting it if it holds an SSH connection."""
-        old_host = self._host_by_id_cache.pop(host_id, None)
-        if old_host is not None and old_host is not replacement and isinstance(old_host, Host):
-            old_host.disconnect()
-        if replacement is not None:
-            self._host_by_id_cache[host_id] = replacement
 
     def _ensure_lima_available(self) -> None:
         """Lazily check that limactl is installed and meets version requirements.
